Updating metadata blocks for CAFE collection
Fixing the spatial resolution fields in both metadata blocks. In the "Metadata About Data Sources"  metadata block, the spatial resolution fields are now part of a compound field, so that depositors more easily see the relationships between the fields.

Also updated how the values of these fields, in both metadata blocks, are displayed on the dataset page.